+++ frontend/src/App.jsx
@@ -1,11 +1,7 @@
-import { useState } from "react";
-import reactLogo from "./assets/react.svg";
-import viteLogo from "/vite.svg";
 import "./App.css";
+import LoginPage from "@/pages/LoginPage.jsx";
 
 function App() {
-  const [count, setCount] = useState(0);
-
   return (
     <>
       <LoginPage />
